validation-service/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
jsonschema==4.20.0

# Message Queue
kafka-python==2.0.2

# Object Storage
minio==7.2.3

# Data Processing
pandas==2.1.4
pyarrow==14.0.2
openpyxl==3.1.2

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2